+++ habana_lazy/ir.h
@@ -571,7 +571,7 @@ class Node {
   std::shared_ptr<std::string> m_scope;
   uint64_t m_id;
   bool deterministic = 0;
-  std::unordered_map<uint32_t, uint32_t> m_pt_vec_to_input_ival;
+  std::unordered_map<size_t, size_t> m_pt_vec_to_input_ival;
   std::string module_name = std::string();
 };
